// emdash/emdash-env.d.ts
// Generated by EmDash on dev server start
// Do not edit manually

/// <reference types="emdash/locals" />

import type { BylineSummary, ContentBylineCredit, TaxonomyTerm, PortableTextBlock } from "emdash";

export interface Event {
  id: string;
  slug: string | null;
  status: string;
  title: string;
  date: string;
  all_day?: boolean;
  start_time?: string;
  end_time?: string;
  location?: string;
  address?: string;
  player_cap?: number;
  levels?: string;
  cancelled?: boolean;
  intro?: string;
  special_note?: string;
  gamemaster?: string;
  scenarios?: { "name": string; "system"?: "Pathfinder" | "Starfinder" | null; "edition"?: "1E" | "2E" | null; "type"?: "Scenario" | "Quest" | "Adventure" | "Bounty" | "Module" | "One-Shot" | null; "levels"?: string | null; "start_time"?: string | null; "end_time"?: string | null; "player_cap"?: number | null; "gamemaster"?: string | null; "repeatable"?: boolean | null; "pregens"?: boolean | null; "cancelled"?: boolean | null; "signup_url"?: string | null }[];
  body?: PortableTextBlock[];
  createdAt: Date;
  updatedAt: Date;
  publishedAt: Date | null;
  byline?: BylineSummary | null;
  bylines?: ContentBylineCredit[];
  terms?: Record<string, TaxonomyTerm[]>;
}

export interface Gamemaster {
  id: string;
  slug: string | null;
  status: string;
  title: string;
  date?: string;
  first_name: string;
  last_initial: string;
  organized_play_number?: number;
  games?: ("Pathfinder" | "Starfinder")[];
  venture_officer?: string;
  bio?: PortableTextBlock[];
  createdAt: Date;
  updatedAt: Date;
  publishedAt: Date | null;
  byline?: BylineSummary | null;
  bylines?: ContentBylineCredit[];
  terms?: Record<string, TaxonomyTerm[]>;
}

export interface New {
  id: string;
  slug: string | null;
  status: string;
  title: string;
  date: string;
  body: PortableTextBlock[];
  createdAt: Date;
  updatedAt: Date;
  publishedAt: Date | null;
  byline?: BylineSummary | null;
  bylines?: ContentBylineCredit[];
  terms?: Record<string, TaxonomyTerm[]>;
}

declare module "emdash" {
  interface EmDashCollections {
    events: Event;
    gamemasters: Gamemaster;
    news: New;
  }
}